refactor(command): rename to singular @shared/command and move renderer to features/command (naming-conventions §4.8–§4.9)
