+++ src/renderer/features/command/__tests__/CommandProvider.test.tsx
@@ -1,4 +1,4 @@
-import type { CommandId } from '@shared/commands'
+import type { CommandId } from '@shared/command'
 import { cleanup, fireEvent, render, screen } from '@testing-library/react'
 import type { ReactNode } from 'react'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
