+++ src/renderer/features/command/__tests__/menus.test.tsx
