+++ src/renderer/pages/agents/AgentNavbar.tsx
@@ -1,8 +1,8 @@
 import { usePreference } from '@data/hooks/usePreference'
-import { useCommandHandler } from '@renderer/commands'
 import { Navbar, NavbarCenter, NavbarLeft, NavbarRight } from '@renderer/components/app/Navbar'
 import NavbarIcon from '@renderer/components/NavbarIcon'
 import SearchPopup from '@renderer/components/Popups/SearchPopup'
+import { useCommandHandler } from '@renderer/features/command'
 import { Tooltip } from 'antd'
 import { t } from 'i18next'
 import { Menu, PanelLeftClose, PanelRightClose, Search } from 'lucide-react'
