+++ src/renderer/pages/agents/components/AgentChatNavbar/index.tsx
@@ -1,6 +1,6 @@
-import { useCommandHandler } from '@renderer/commands'
 import { NavbarHeader } from '@renderer/components/app/Navbar'
 import SearchPopup from '@renderer/components/Popups/SearchPopup'
+import { useCommandHandler } from '@renderer/features/command'
 import { cn } from '@renderer/utils'
 import type { AgentEntity } from '@shared/data/types/agent'
