+++ src/renderer/pages/home/Inputbar/Inputbar.tsx
@@ -1,9 +1,9 @@
 import { cacheService } from '@data/CacheService'
 import { usePreference } from '@data/hooks/usePreference'
 import { loggerService } from '@logger'
-import { useCommandHandler } from '@renderer/commands'
 import { isGenerateImageModel, isGenerateImageModels, isVisionModel, isVisionModels } from '@renderer/config/models'
 import { useCache } from '@renderer/data/hooks/useCache'
+import { useCommandHandler } from '@renderer/features/command'
 import { useAssistant } from '@renderer/hooks/useAssistant'
 import { useInputText } from '@renderer/hooks/useInputText'
 import { useKnowledgeBases } from '@renderer/hooks/useKnowledgeBase'
