+++ src/renderer/pages/home/components/ChatNavBar/index.tsx
@@ -1,8 +1,8 @@
 import { Tooltip } from '@cherrystudio/ui'
 import { usePreference } from '@data/hooks/usePreference'
-import { useCommandHandler } from '@renderer/commands'
 import { NavbarHeader } from '@renderer/components/app/Navbar'
 import SearchPopup from '@renderer/components/Popups/SearchPopup'
+import { useCommandHandler } from '@renderer/features/command'
 import { useNavbarPosition } from '@renderer/hooks/useNavbar'
 import { t } from 'i18next'
 import { PanelLeftClose, PanelRightClose } from 'lucide-react'
