+++ src/renderer/windows/subWindow/SubWindowApp.tsx
@@ -1,7 +1,6 @@
 import '@renderer/databases'
 
 import { preferenceService } from '@data/PreferenceService'
-import { CommandProvider, ContextKeyProvider } from '@renderer/commands'
 import TopViewContainer from '@renderer/components/TopView'
 import AntdProvider from '@renderer/context/AntdProvider'
 import { CodeStyleProvider } from '@renderer/context/CodeStyleProvider'
@@ -9,6 +8,7 @@ import { NotificationProvider } from '@renderer/context/NotificationProvider'
 import StyleSheetManager from '@renderer/context/StyleSheetManager'
 import { TabsProvider } from '@renderer/context/TabsContext'
 import { ThemeProvider } from '@renderer/context/ThemeProvider'
+import { CommandProvider, ContextKeyProvider } from '@renderer/features/command'
 import store, { persistor } from '@renderer/store'
 import { SubWindowAppShell } from '@renderer/windows/subWindow/SubWindowAppShell'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
