+++ src/shared/command/__tests__/keybindings.test.ts
