+++ src/shared/command/__tests__/menus.test.ts
