+++ src/shared/command/definitions.ts
